Разбил методы по правилам ES2015
